(build) synch vendored single-header implementations
FossilOrigin-Name: 81ff94f473833b0f80b18da6fbda0aff24df18f856cabe7ece4d216ef5f56c8b

+++ src/tools/aloadimage/imgload.c
@@ -279,8 +279,8 @@ bool imgload_poll(struct img_state* tgt)
 		tgt->out = NULL;
 		snprintf((char*)tgt->msg, sizeof(tgt->msg),
 			"(%zuMiB>%zuMiB)",
-			(out_sz+1) / (1024 * 1024),
-			(tgt->buf_lim+1) / (1024 * 1024)
+			(size_t)((out_sz+1) / (1024 * 1024)),
+			(size_t)((tgt->buf_lim+1) / (1024 * 1024))
 		);
 		mark_finished(tgt, true);
 		return true;
